cli: split into files
