+++ cli/pyproject.toml
@@ -25,7 +25,7 @@ classifiers = [
 ]
 
 [project.scripts]
-meirim = 'src.meirim:cli'
+meirim = 'src.cli:entry_point'
 
 [project.urls]
 "Homepage" = "https://github.com/meirim-org/meirim"
